refactor extensions and identity packages
