+++ pkg/util/file/file.go
@@ -20,7 +20,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/sigstore/rekor-monitor/pkg/rekor"
+	"github.com/sigstore/rekor-monitor/pkg/identity"
 	"github.com/sigstore/rekor/pkg/util"
 )
 
@@ -118,7 +118,7 @@ func DeleteOldCheckpoints(logInfoFile string) error {
 }
 
 // WriteIdentity writes an identity found in the log to a file
-func WriteIdentity(idFile string, idEntry rekor.IdentityEntry) error {
+func WriteIdentity(idFile string, idEntry identity.RekorLogEntry) error {
 	file, err := os.OpenFile(idFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
 		return fmt.Errorf("failed to open identities file: %w", err)
